+++ src/frontend/src/pages/part/pricing/PricingPanel.tsx
@@ -1,0 +1,27 @@
+import { Accordion } from '@mantine/core';
+import { ReactNode } from 'react';
+
+import { StylishText } from '../../../components/items/StylishText';
+
+export default function PricingPanel({
+  content,
+  label,
+  title,
+  visible
+}: {
+  content: ReactNode;
+  label: string;
+  title: string;
+  visible: boolean;
+}): ReactNode {
+  return (
+    visible && (
+      <Accordion.Item value={label}>
+        <Accordion.Control>
+          <StylishText size="lg">{title}</StylishText>
+        </Accordion.Control>
+        <Accordion.Panel>{content}</Accordion.Panel>
+      </Accordion.Item>
+    )
+  );
+}
